+++ rustysynth/src/volume_envelope.rs
@@ -20,7 +20,7 @@ pub(crate) struct VolumeEnvelope {
     sustain_level: f32,
     release_level: f32,
 
-    processed_sample_count: i32,
+    processed_sample_count: usize,
     stage: i32,
     value: f32,
 
@@ -81,7 +81,7 @@ impl VolumeEnvelope {
         self.release_level = self.value;
     }
 
-    pub(crate) fn process(&mut self, sample_count: i32) -> bool {
+    pub(crate) fn process(&mut self, sample_count: usize) -> bool {
         self.processed_sample_count += sample_count;
 
         let current_time = self.processed_sample_count as f64 / self.sample_rate as f64;
